Add a naive CSV splitter
It is naive because it does not handle escape characters. It does cope with single pairs of quotes, however, which is a good starting point.

Note that this change also includes some changes to the CSV parser tests. Quotes will temporarily be permitted inside a value. In other words, the only characters that require quoting are commas themselves. This is likely not expected in the long-term, but suffices for now.