Simplify splitting logic
Using one if-else-if statement was a micro-optimisation that tangled the logic around encountered characters. The majority of characters in a CSV will not be one of the special characters we care about, so will go through all the tests.

By using separate conditions, the logic for each character can be packaged much more neatly, albeit still not entirely decoupled.